backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
celery==5.4.0
redis==5.1.0
minio==7.2.9
faster-whisper==1.0.3
ffmpeg-python==0.2.0
numpy==1.26.4
websockets==13.1
aiofiles==24.1.0
httpx==0.27.2
python-dotenv==1.0.1
requests==2.32.3
